# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste prático - Backend",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/tarefas": {
            "get": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Listar tarefas",
                "description": "Retornar todas as tarefas cadastradas",
                "operationId": "listar_tarefas",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Cadastrar nova tarefa",
                "operationId": "nova_tarefa",
                "requestBody": {
                    "description": "Campos para cadastros",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "descricao",
                                    "data_conclusao"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "titulo",
                                        "format": "text",
                                        "example": "Revisão do carro"
                                    },
                                    "descricao": {
                                        "type": "descricao",
                                        "format": "text",
                                        "example": "Levar o carro na concessionária para realizar a revisão."
                                    },
                                    "data_conclusao": {
                                        "type": "data_conclusao",
                                        "format": "date",
                                        "example": "2023-03-19"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    },
                    "417": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tarefas/{id}": {
            "get": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Detalhar tarefa",
                "description": "Retornar informação de uma tarefa",
                "operationId": "get_tarefa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Alterar detalhe da tarefa",
                "operationId": "editar_tarefa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Campos para cadastros",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "descricao",
                                    "data_conclusao"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "titulo",
                                        "format": "text",
                                        "example": "Revisão do carro"
                                    },
                                    "descricao": {
                                        "type": "descricao",
                                        "format": "text",
                                        "example": "Levar o carro na concessionária para realizar a revisão."
                                    },
                                    "data_conclusao": {
                                        "type": "data_conclusao",
                                        "format": "date",
                                        "example": "2023-03-19"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    },
                    "417": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Excluir tarefa especifica",
                "operationId": "excluir_tarefa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth"
            }
        }
    ]
}